Added dao impl and bug fix
