+++ src/main/java/com/softserve/betterlearningroom/controller/MaterialController.java
@@ -1,5 +1,8 @@
 package com.softserve.betterlearningroom.controller;
 
+import com.softserve.betterlearningroom.dto.MaterialDTO;
+import com.softserve.betterlearningroom.service.MaterialService;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -11,11 +14,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.softserve.betterlearningroom.dto.MaterialDTO;
-import com.softserve.betterlearningroom.service.MaterialService;
-
-import lombok.RequiredArgsConstructor;
-
 import java.util.List;
 
 @RestController
@@ -26,27 +24,27 @@ public class MaterialController {
     private final MaterialService materialService;
 
     @GetMapping("{id}")
-    public ResponseEntity<MaterialDTO> findMaterialById(@PathVariable Long id){
+    public ResponseEntity<MaterialDTO> findMaterialById(@PathVariable Long id) {
         return ResponseEntity.ok().body(materialService.getMaterialById(id));
     }
 
     @GetMapping
-    public ResponseEntity<List<MaterialDTO>> findAllMaterials(@PathVariable Long classroomId){
+    public ResponseEntity<List<MaterialDTO>> findAllMaterials(@PathVariable Long classroomId) {
         return ResponseEntity.ok().body(materialService.getMaterialsByClassroom(classroomId));
     }
 
     @PostMapping
-    public ResponseEntity<MaterialDTO> createMaterial(@RequestBody MaterialDTO material, @PathVariable Long topicId){
+    public ResponseEntity<MaterialDTO> createMaterial(@RequestBody MaterialDTO material, @PathVariable Long topicId) {
         return new ResponseEntity<>(materialService.addMaterial(material, topicId), HttpStatus.CREATED);
     }
 
     @PutMapping("{id}")
-    public ResponseEntity<MaterialDTO> updateMaterial(@RequestBody MaterialDTO material){
+    public ResponseEntity<MaterialDTO> updateMaterial(@RequestBody MaterialDTO material) {
         return new ResponseEntity<>(materialService.updateMaterial(material), HttpStatus.ACCEPTED);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<?> deleteMaterial(@RequestBody MaterialDTO material){
+    public ResponseEntity<?> deleteMaterial(@RequestBody MaterialDTO material) {
         materialService.updateMaterial(material);
         return new ResponseEntity<>(HttpStatus.OK);
     }
