+++ src/main/java/com/softserve/betterlearningroom/dao/extractor/MaterialRowMapper.java
@@ -24,28 +24,18 @@ public class MaterialRowMapper implements ResultSetExtractor<List<Material>> {
     public List<Material> extractData(ResultSet rs) throws SQLException, DataAccessException {
         Map<Long, Material> materialMap = new HashMap<>();
         while (rs.next()) {
-            Long materialId = rs.getLong("materialid");
+            Long materialId = rs.getLong("material_id");
             Material material = materialMap.get(materialId);
             MaterialType type = valueOf(rs.getString("materialType"));
             if (material == null){
-                switch (type){
-                    case TEST:
-                        material = new Test();
-                        break;
-                    case TASK:
-                        material = new Task();
-                        break;
-                    default:
-                        material = new Material();
-                        break;
-                }
+                material = new Material();
+
                 String className = getClass().getSimpleName() + type.name();
-
-
+                material.setDueDate(rs.getTimestamp("duedate").toLocalDateTime());
                 material.setTask(rs.getString("task"));
                 material.setUrl(rs.getString("testUrl"));
-                material.setStartDate(rs.getDate("startdate").toLocalDate().atStartOfDay());
-                material.setDueDate(rs.getDate("duedate").toLocalDate().atStartOfDay());
+                material.setStartDate(rs.getTimestamp("startdate").toLocalDateTime());
+                material.setDueDate(rs.getTimestamp("duedate").toLocalDateTime());
                 material.setMaxScore(rs.getByte("maxScore"));
                 material.setId(materialId);
                 material.setText(rs.getString("materialtext"));
