+++ src/main/java/com/softserve/betterlearningroom/entity/Material.java
@@ -13,13 +13,16 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @ToString
-@EqualsAndHashCode
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
 public class Material {
 
+    @EqualsAndHashCode.Include
     private Long id;
 
+    @EqualsAndHashCode.Include
     private MaterialType materialType;
 
+    @EqualsAndHashCode.Include
     private String title;
 
     private String text;
