+++ src/main/java/com/softserve/betterlearningroom/entity/MaterialType.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 public enum MaterialType {
     TASK("TASK"), TEST("TEST"), QUESTIONS("QUESTIONS"), MATERIAL("MATERIAL");
 
-    private String name;
+    private final String name;
 
     MaterialType(String name) {
         this.name = name;
