+++ src/main/resources/classroomQuery.properties
@@ -1,4 +1,5 @@
 getClassroomById = SELECT classroom_id, user_id, title, session, description, code FROM classrooms WHERE classroom_id = :classroomId
+getClassroomByOwnerId = SELECT classroom_id, user_id, title, session, description, code FROM classrooms WHERE user_id = :userId
 getClassroomTeachers = SELECT id, firstname, lastname, email, password, enabled FROM users JOIN classrooms ON users.id = classrooms.user_id WHERE classroom_id = :classroomId
 getClassroomStudents = SELECT id, firstname, lastname, email, password, enabled FROM users JOIN classrooms ON users.id = classrooms.user_id WHERE classroom_id = :classroomId
 getClassroomOwnerById = SELECT id, firstname, lastname, email, password, enabled FROM users JOIN classrooms ON users.id = classrooms.user_id WHERE classroom_id = :classroomId
@@ -6,7 +7,7 @@ getClassroomsByTeacher = SELECT classrooms.classroom_id, classrooms.user_id, tit
   JOIN teachers_classrooms ON teachers_classrooms.classroom_id = classrooms.classroom_id WHERE teachers_classrooms.user_id = :userId
 getClassroomsByStudent = SELECT classrooms.classroom_id, classrooms.user_id, title, session, description, code FROM classrooms \
   JOIN students_classrooms ON students_classrooms.classroom_id = classrooms.classroom_id WHERE students_classrooms.user_id = :userId
-createClassroom = INSERT INTO classrooms (classroom_id, user_id, title, session, description, code) VALUES (:classroomId, :userId, :title, :session, :description, random()::text)
+createClassroom = INSERT INTO classrooms (user_id, title, session, description, code) VALUES (:userId, :title, :session, :description, MD5(random()::text))
 removeClassroom = DELETE FROM classrooms WHERE classroom_id = :classroomId
 getClassroomByCode = SELECT classroom_id, user_id, title, session, description, code FROM classrooms WHERE code = :code
 joinClassroomAsStudent = INSERT INTO students_classrooms (classroom_id, user_id) VALUES (:classroomId, :userId)
