+++ src/main/resources/db/materials/materialQuery.properties
@@ -1,19 +1,19 @@
-get.all.materials = SELECT mtr.materialid, mtr.materialtext, mtr.startdate, \
+get.all.materials = SELECT mtr.material_id, mtr.materialtext, mtr.startdate, \
   mtr.duedate, mtr.task, mtr.materialtype, mtr.maxscore, \
   mtr.testurl, mtr.title, mtr.topicid, \
   lin.linkid, lin.url, lin.linktext, q.questionid, q.question \
   FROM materials as mtr \
-  FULL JOIN links AS lin ON mtr.materialid=lin.materialid \
-  FULL JOIN questions as q ON mtr.materialid=q.materialid \
+  FULL JOIN links AS lin ON mtr.material_id=lin.materialid \
+  FULL JOIN questions as q ON mtr.material_id=q.materialid \
   WHERE mtr.topicid=:topicid
-get.by.id.material = SELECT mtr.materialid, mtr.materialtext, mtr.startdate, \
+get.by.id.material = SELECT mtr.material_id, mtr.materialtext, mtr.startdate, \
   mtr.duedate, mtr.task, mtr.materialtype, mtr.maxscore, \
   mtr.testurl, mtr.title, lin.linkid, lin.url, lin.linktext, \
   q.questionid, q.question \
   FROM materials AS mtr \
-  FULL JOIN links as lin ON mtr.materialid=lin.materialid \
-  FULL JOIN questions as q ON mtr.materialid=q.materialid \
-  WHERE q.materialid=:materialid OR lin.materialid=:materialid
+  FULL JOIN links as lin ON mtr.material_id=lin.materialid \
+  FULL JOIN questions as q ON mtr.material_id=q.materialid \
+  WHERE mtr.material_id=:materialid
 add.new.material = INSERT INTO materials (title, materialtext, startdate, duedate, task, materialtype, maxscore, testurl, topicid) \
   VALUES (:title, :materialtext, :startdate, :duedate, :task, :materialtype, :maxscore, :testurl, :topicid)
 update.material = UPDATE materials SET title=:title, materialtext=:materialtext, startdate=:startdate, duedate=:duedate, \
